[REFACTOR] Gateway 구조 리팩토링
[REFACTOR] Gateway 구조 리팩토링
